requirements.txt
altair
django-extensions
goerr